refactor: separate API cache updates and utilities
