+++ frontend/src/api/cache/updates/categories.ts
