+++ frontend/src/api/cache/utils/referenceLookup.ts
