+++ frontend/src/api/transactions/index.ts
@@ -22,10 +22,6 @@ export {
 } from '@/api/transactions/requests';
 
 export {
-  invalidateTransactionAccountData,
-} from '@/api/cache/transactions';
-
-export {
   useCreateTransaction,
   useDeleteTransaction,
   useInfiniteTransactions,
